Revert move of namespace list.
